refactor: 초기 데이터 설정 디렉토리 수정 (#98)
