+++ backend/src/main/java/ku_rum/backend/global/dataInsert/CategoryInitializer.java
@@ -1,4 +1,4 @@
-package ku_rum.backend.global.dataInsert;
+package ku_rum.backend.global.dataInit;
 
 import ku_rum.backend.domain.category.domain.Category;
 import ku_rum.backend.domain.category.domain.CategoryType;
